# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AIRBNB API",
        "version": "1.0.0",
        "description": "documentation for airbnb API",
        "contact": {
            "name": "Team 1 Alterra",
            "email": "ezabintangramadhan@gmail.com"
        },
        "license": {
            "name": "APACHE 2.0"
        }
    },
    "servers": [
        {
            "description": "SwaggerHub API Auto Mocking",
            "url": "https://virtserver.swaggerhub.com/ezabintangramadhan/Airbnb-API/1.0.0"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login user",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "description": "Username user",
                                        "example": "user123"
                                    },
                                    "password": {
                                        "type": "string",
                                        "description": "passwword user",
                                        "example": "passwword123"
                                    }
                                },
                                "required": [
                                    "username",
                                    "passwword"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Login"
                                },
                                "examples": {
                                    "success": {
                                        "description": "Login successful",
                                        "value":{
                                            "id": 1,
                                            "name":"user123",
                                            "token":"agaTwefhwoefh12314"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400":{
                        "description": "Login failed",
                        "content": {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "description":{
                                            "type": "string"
                                        }
                                    }
                                },
                                "examples": {
                                    "failed": {
                                        "description": "Login failed",
                                        "value":{
                                            "status":"login failed",
                                            "description":"wrong username/password"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "user not found. please input correctly.",
                    "content": {
                    "application/json":{
                        "schema":{
                            "type": "object",
                            "properties": {
                                "message":{
                                    "type": "string"
                                }
                            },
                            "example":{
                                "message":"user not found. please input correctly."
                            }
                        }
                    }
                }
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/users": {
            "post":{
                "tags": ["users"],
                "summary": "create new user",
                "description": "",
                "requestBody":{
                    "required": true,
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/InsertUser"
                            },
                            "example": {
                                "name":"tukimin",
                                "email":"tukimin@mail.com",
                                "password":"tukimin123",
                                "phone":"08220321",
                                "role":"user"
                            }
                        }
                    }
                },
                "responses": {
                    "201":{
                       "$ref": "#/components/responses/201"
                    },
                    "400":{
                        "$ref": "#/components/responses/400"
                    },
                    "404":{
                        "$ref": "#/components/responses/404"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "get": {
                "security": [
                {
                    "JWTAuth":[]
                }
                ],
                "tags": [
                    "users"
                ],
                "summary": "see profile user",
                "description": "see profile user by token id",
                "responses": {
                    "200": {
                        "description": "Success get profile user",
                        "content": {
                            "application/json":{
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name":{
                                            "type": "string",
                                            "example": "tukimin"
                                        },
                                        "email":{
                                            "type": "string",
                                            "example":"tukimin@mail.com"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update existing user",
                "description": "update existing user by id",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUser"
                            },
                            "examples": {
                                "updateUser":{
                                    "value":{
                                        "name":"Eza",
                                        "email":"eza@mail.com",
                                        "password":"eza123",
                                        "phone":"0822232"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful update",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref": "#/components/responses/200"
                                },
                                "examples":{
                                    "successfulUpdate":{
                                        "value":{
                                            "message":"success update user"
                                        }
                                    }
                               }
                           }
                        }
                    },
                    "400":{
                        "description": "wrong input",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "$ref": "#/components/responses/400"
                                },
                                "examples":{
                                    "failedUpdate":{
                                        "value":{
                                            "message":"failed update. input your changes correctly"
                                        }
                                    }
                               }
                            }
                        }
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "delete":{
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": [
                    "users"
                ],
                "summary": "delete existing user",
                "description": "delete existing user by id",
                "responses": {
                    "200":{
                        "description": "successful delete user",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "$ref": "#/components/responses/200"
                                },
                                "examples":{
                                    "successfulUpdate":{
                                        "value":{
                                            "message":"success delete user"
                                        }
                                    }
                               }
                           }
                        }
                    },
                    "404":{
                        "description": "user not found",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "$ref": "#/components/responses/404"
                                },
                                "examples":{
                                    "failedDelete":{
                                        "value":{
                                            "message":"failed delete. user not found"
                                        }
                                    }
                               }
                            }
                        }
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/homestays":{
            "get":{
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": [
                    "homestays"
                ],
                "summary": "get all homestay in user mode",
                "description": "get all homestay when user in user mode",
                "responses": {
                    "200":{
                        "description": "Success get all homestay",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf":[
                                        {
                                            "$ref": "#/components/schemas/ResponsesWithData"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message":{
                                                    "type": "string",
                                                    "example":"success get all homestay"
                                                },
                                                "data":{
                                                    "type": "object",
                                                    "properties": {
                                                        "homestayID" :{
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "homestayName":{
                                                            "type": "string",
                                                            "example":"puncak cottage"
                                                        },
                                                        "address":{
                                                            "type": "string",
                                                            "example":"Bandung"
                                                        },
                                                        "pricePerNight":{
                                                            "type": "integer",
                                                            "example":120000
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404":{
                        "$ref": "#/components/responses/404"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/homestays/{id}":{
            "get":{
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": ["homestays"],
                "summary": "get homestay by id",
                "description": "get homestay using id at path parameter. only return 1 data",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/idHomestay"
                    }
                ],
                "responses": {
                    "200":{
                        "description": "Success get homestay",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf":[
                                        {
                                            "$ref": "#/components/schemas/ResponsesWithData"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message":{
                                                    "type": "string",
                                                    "example":"success get homestay"
                                                },
                                                "data":{
                                                    "type": "object",
                                                    "properties": {
                                                        "homestayID" :{
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "homestayName":{
                                                            "type": "string",
                                                            "example":"puncak cottage"
                                                        },
                                                        "address":{
                                                            "type": "string",
                                                            "example":"Bandung"
                                                        },
                                                        "pricePerNight":{
                                                            "type": "integer",
                                                            "example":120000
                                                        },
                                                        "description":{
                                                            "type": "string",
                                                            "example":"1 queen size bed, 1 toilet, TV, wifi, etc"
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404":{
                        "$ref": "#/components/responses/404"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/homestays/host":{
            "get":{
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": ["homestays"],
                "summary": "get all homestay in host mode",
                "description": "get all homestay when user in host mode",
                "responses": {
                    "200":{
                        "description": "Success get all homestay",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf":[
                                        {
                                            "$ref": "#/components/schemas/ResponsesWithData"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message":{
                                                    "type": "string",
                                                    "example":"success get all homestay"
                                                },
                                                "data":{
                                                    "type": "object",
                                                    "properties": {
                                                        "homestayID" :{
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "homestayName":{
                                                            "type": "string",
                                                            "example":"puncak cottage"
                                                        },
                                                        "address":{
                                                            "type": "string",
                                                            "example":"Bandung"
                                                        },
                                                        "pricePerNight":{
                                                            "type": "integer",
                                                            "example":120000
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404":{
                        "$ref": "#/components/responses/404"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            },
            "post":{
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": ["homestays"],
                "summary": "create new homestay",
                "description": "create homestays for rent",
                "requestBody":{
                    "required": true,
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/CreateHomestay"
                            },
                            "example":{
                                "homestayName":"puncak dieng",
                                "address":"Bandung",
                                "description":"1 queen size bed, 1 toilet, TV, wifi, etc",
                                "pricePerNight":145000,
                                "images1":"url images1",
                                "images2":"url images2",
                                "images3":"url images3"
                            }
                        }
                    }
                },
                "responses": {
                    "201":{
                        "description": "success add homestay",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf":[
                                        {
                                            "$ref": "#/components/responses/201"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message":{
                                                    "type": "string",
                                                    "example":"success add homestay"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400":{
                        "$ref": "#/components/responses/400"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
        "/homestays/host/myHomestay":{
            "get":{
                "security": [
                    {
                        "JWTAuth":[]
                    }
                ],
                "tags": ["homestays"],
                "description": "get all user's homestay",
                "responses": {
                    "200":{
                        "description": "Success get all homestay",
                        "content": {
                            "application/json":{
                                "schema":{
                                    "allOf":[
                                        {
                                            "$ref": "#/components/schemas/ResponsesWithData"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "message":{
                                                    "type": "string",
                                                    "example":"success get all homestay"
                                                },
                                                "data":{
                                                    "type": "object",
                                                    "properties": {
                                                        "homestayID" :{
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "homestayName":{
                                                            "type": "string",
                                                            "example":"puncak cottage"
                                                        },
                                                        "address":{
                                                            "type": "string",
                                                            "example":"Bandung"
                                                        },
                                                        "pricePerNight":{
                                                            "type": "integer",
                                                            "example":120000
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404":{
                        "$ref": "#/components/responses/404"
                    },
                    "500":{
                        "$ref": "#/components/responses/500"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWTAuth": {
                "description": "token for admin",
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }

        },
...
        "schemas": {
            "Login": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
            "InsertUser":{
                "type": "object",
                "properties": {
                    "name":{
                        "type": "string"
                    },
                    "email":{
                        "type": "string"
                    },
                    "password":{
                        "type": "string"
                    },
                    "phone":{
                        "type": "string"
                    },
                    "role":{
                        "type": "string"
                    }
                }
            },
            "ResponsesWithData": {
                "type": "object",
                "properties": {
                    "message":{
                        "type": "string"
                    },
                    "data":{
                        "type": "object"
                    }
                }
            },
...
        },
        "responses": {
...
            "201":{
                "description": "success create data",
                "content": {
                    "application/json":{
                        "schema":{
                            "type": "object",
                            "properties": {
                                "message":{
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "400":{
                "description": "bad request from user",
                "content": {
                    "application/json":{
                        "schema":{
                            "type": "object",
                            "properties": {
                                "message":{
                                    "type": "string"
                                }
                            },
                            "example":{
                                "message":"Process failed. Please check your input"
                            }
                        }
                    }
                }
            },
...
            "500":{
                "description": "error procces on server",
                "content": {
                    "application/json":{
                        "schema":{
                            "type": "object",
                            "properties": {
                                "message":{
                                    "type": "string"
                                }
                            },
                            "example":{
                                "message": "procces failed. error internal server"
                            }
                        }
                    }
                }
            }
        }
    }
}